% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{juvenile_theme}
\alias{juvenile_theme}
\title{Theme for juvenile herring plots}
\usage{
juvenile_theme()
}
\value{
A customized \code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\description{
Customized theme for juvenile herring plots with ggplot
}
