% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_rbr}
\alias{load_rbr}
\title{Load RBR data.}
\usage{
load_rbr(where, ts_info = NULL, quiet = TRUE, ...)
}
\arguments{
\item{where}{Location of the input data file.}

\item{ts_info}{Tibble. Date and transect information. Default NULL.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}

\item{...}{Other arguments passed to \code{\link[oce:read.rsk]{oce::read.rsk()}}.}
}
\value{
tibble
}
\description{
Load data collected by the RBR sensor used in Bongo net casts.
}
\note{
Before loading the data, process the data by hand to identify transect
and station.
}
\examples{
data(ts_data)
load_rbr(where = here::here("data-raw", "Test.rsk"), ts_info = NULL)
}
\references{
\insertAllCited{}
}
\seealso{
oce::read.rsk(), ts_data
}
\concept{load functions}
