% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_rbr}
\alias{load_rbr}
\title{Load RBR data.}
\usage{
load_rbr(where, quiet = TRUE)
}
\arguments{
\item{where}{Location of the input data file.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
tibble
}
\description{
Load data collected by the RBR sensor used in Bongo net casts.
}
\note{
Before loading the data, process the data by hand to identify transect
and station.
}
\examples{
load_rbr(where = "here")
}
\references{
\insertAllCited{}
}
\concept{load functions}
