% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_data}
\alias{ts_data}
\title{Transect and station data.}
\format{
Tibble with rows for dates and columns for attributes.
}
\usage{
ts_data
}
\description{
Table with dates for transects and stations.
}
\examples{
data(ts_data)
ts_data
}
\references{
\insertAllCited{}
}
\concept{data}
\keyword{datasets}
